requirements.txt
openai>=1.0.0
requests
pandas
yfinance
python-dotenv
openpyxl
